3 funktionen ausgelagert, inputcheck fehlt noch
